added traveling music to main level
I went in and added a new sub folder for the music folder. The New folder name is called Background music. This folder contains the traveling music and the Main menu music. For the traveling music. I added an audio listener to the background image and inserted the traveling music to the audio listener.